# swagger.yaml
definitions:
  Activities:
    items:
      $ref: '#/definitions/Activity'
    type: array
    x-go-package: houseduino-be/models
  Activity:
    properties:
      data_inserimento:
        description: |-
          Timestamp of insert
          in: time
        format: date-time
        type: string
        x-go-name: DateInsert
      id:
        description: |-
          Id of Activity value
          in: int64
        format: int64
        type: integer
        x-go-name: Id
      valore:
        description: |-
          Value of Activity
          in: int
        format: int64
        type: integer
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
  Altitude:
    properties:
      data_inserimento:
        description: |-
          Timestamp of insert
          in: time
        format: date-time
        type: string
        x-go-name: DateInsert
      id:
        description: |-
          Id of rain value
          in: int64
        format: int64
        type: integer
        x-go-name: Id
      valore:
        description: |-
          Value of Altitude
          in: int
        format: float
        type: number
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
  Altitudes:
    items:
      $ref: '#/definitions/Altitude'
    type: array
    x-go-package: houseduino-be/models
  CommonError:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
...
  GetActivities:
    properties:
      data:
        $ref: '#/definitions/Activities'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetActivity:
    properties:
      data:
        $ref: '#/definitions/Activity'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetAltitude:
    properties:
      data:
        $ref: '#/definitions/Altitude'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetAltitudes:
    properties:
      data:
        $ref: '#/definitions/Altitudes'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetHumidities:
    properties:
      data:
        $ref: '#/definitions/Humidities'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetHumidity:
    properties:
      data:
        $ref: '#/definitions/Humidity'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetIsActive:
    properties:
      active:
        description: |-
          Message of the response
          in: string
        type: boolean
        x-go-name: Active
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
...
  GetRain:
    properties:
      data:
        $ref: '#/definitions/Rain'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
  GetRains:
    properties:
      data:
        $ref: '#/definitions/Rains'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: houseduino-be/controllers
...
  Humidities:
    items:
      $ref: '#/definitions/Humidity'
    type: array
    x-go-package: houseduino-be/models
  Humidity:
    properties:
      data_inserimento:
        description: |-
          Timestamp of insert
          in: time
        format: date-time
        type: string
        x-go-name: DateInsert
      id:
        description: |-
          Id of rain value
          in: int64
        format: int64
        type: integer
        x-go-name: Id
      valore:
        description: |-
          Value of Humidity
          in: int
        format: float
        type: number
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
...
  Rain:
    properties:
      data_inserimento:
        description: |-
          Timestamp of insert
          in: time
        format: date-time
        type: string
        x-go-name: DateInsert
      id:
        description: |-
          Id of rain value
          in: int64
        format: int64
        type: integer
        x-go-name: Id
      valore:
        description: |-
          Value of rain
          in: int
        format: int64
        type: integer
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
  Rains:
    items:
      $ref: '#/definitions/Rain'
    type: array
    x-go-package: houseduino-be/models
  ReqAddActivity:
    properties:
      valore:
        description: |-
          Value of the Activity
          in: int
        format: int64
        type: integer
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
  ReqAddAltitude:
    properties:
      valore:
        description: |-
          Value of the Altitude
          in: int
        format: float
        type: number
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
  ReqAddHumidity:
    properties:
      valore:
        description: |-
          Value of the Humidity
          in: int
        format: float
        type: number
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
...
  ReqAddRain:
    properties:
      valore:
        description: |-
          Value of the rain
          in: int
        format: int64
        type: integer
        x-go-name: Value
    type: object
    x-go-package: houseduino-be/models
...
info:
  title: 'Houseduino Api:'
  version: 0.1.1
paths:
  /activity/all:
    get:
      description: Get Activity list
      operationId: activityAll
      responses:
        "200":
          description: GetActivities
          schema:
            $ref: '#/definitions/GetActivities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - activity
  /activity/insert:
    post:
      description: Create a new Activity value
      operationId: addActivity
      parameters:
      - description: 'name: body'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReqAddActivity'
        x-go-name: Body
      responses:
        "200":
          description: GetActivity
          schema:
            $ref: '#/definitions/GetActivity'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - activity
  /activity/isactive:
    get:
      description: Get if sensor is online or offline
      operationId: activityIsActive
      responses:
        "200":
          description: GetIsActive
          schema:
            $ref: '#/definitions/GetIsActive'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - activity
  /activity/lasthour:
    get:
      description: Get list of last hour of Activity .... or the last value inserted
      operationId: activityLasthour
      responses:
        "200":
          description: GetActivities
          schema:
            $ref: '#/definitions/GetActivities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - activity
  /altitude/all:
    get:
      description: Get Altitude list
      operationId: altitudeAll
      responses:
        "200":
          description: GetAltitudes
          schema:
            $ref: '#/definitions/GetAltitudes'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - altitude
  /altitude/insert:
    post:
      description: Create a new altitude value
      operationId: addAltitude
      parameters:
      - description: 'name: body'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReqAddAltitude'
        x-go-name: Body
      responses:
        "200":
          description: GetAltitude
          schema:
            $ref: '#/definitions/GetAltitude'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - altitude
  /altitude/lasthour:
    get:
      description: Get list of last hour of altitude values .... or the last value
        inserted
      operationId: altitudeLastHour
      responses:
        "200":
          description: GetAltitudes
          schema:
            $ref: '#/definitions/GetAltitudes'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - altitude
  /altitude/showdata/{recordNumber}:
    get:
      description: Get list of recordNumber of altitude values
      operationId: altitudeShowdata
      parameters:
      - description: maximum numnber of results to return
        format: int32
        in: path
        name: recordNumber
        required: true
        type: integer
      responses:
        "200":
          description: GetAltitudes
          schema:
            $ref: '#/definitions/GetAltitudes'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - altitude
  /humidity/all:
    get:
      description: Get humidity list
      operationId: humidityAll
      responses:
        "200":
          description: GetHumidities
          schema:
            $ref: '#/definitions/GetHumidities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - humidity
  /humidity/insert:
    post:
      description: Create a new humidity value
      operationId: addHumidity
      parameters:
      - description: 'name: body'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReqAddHumidity'
        x-go-name: Body
      responses:
        "200":
          description: GetHumidity
          schema:
            $ref: '#/definitions/GetHumidity'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - humidity
  /humidity/lasthour:
    get:
      description: Get list of last hour of humidity values .... or the last value
        inserted
      operationId: humiditylastHour
      responses:
        "200":
          description: GetHumidities
          schema:
            $ref: '#/definitions/GetHumidities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - humidity
  /humidity/showdata/{recordNumber}:
    get:
      description: Get list of recordNumber of humidity values
      operationId: humidityShowdata
      parameters:
      - description: maximum numnber of results to return
        format: int32
        in: path
        name: recordNumber
        required: true
        type: integer
      responses:
        "200":
          description: GetHumidities
          schema:
            $ref: '#/definitions/GetHumidities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - humidity
...
  /rain/all:
    get:
      description: Get rains list
      operationId: rainAll
      responses:
        "200":
          description: GetRains
          schema:
            $ref: '#/definitions/GetRains'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - rain
  /rain/insert:
    post:
      description: Create a new rain value
      operationId: addRain
      parameters:
      - description: 'name: body'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReqAddRain'
        x-go-name: Body
      responses:
        "200":
          description: GetRain
          schema:
            $ref: '#/definitions/GetRain'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - rain
  /rain/lasthour:
    get:
      description: Get list of last hour of rain .... or the last value inserted
      operationId: rainLastHour
      responses:
        "200":
          description: GetRains
          schema:
            $ref: '#/definitions/GetRains'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - rain
  /rain/showdata/{recordNumber}:
    get:
      description: Get list of recordNumber of rain values
      operationId: rainShowdata
      parameters:
      - description: maximum numnber of results to return
        format: int32
        in: path
        name: recordNumber
        required: true
        type: integer
      responses:
        "200":
          description: GetRains
          schema:
            $ref: '#/definitions/GetRains'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - rain
...
produces:
- application/json
schemes:
- http
swagger: "2.0"